use getAll
