+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/program/notification/ProgramNotificationInstanceService.java
@@ -47,4 +47,6 @@ public interface ProgramNotificationInstanceService {
   Long countProgramNotificationInstances(ProgramNotificationInstanceParam params);
 
   void validateQueryParameters(ProgramNotificationInstanceParam param);
+
+  List<ProgramNotificationInstance> getAll();
 }
